fixbug search students in class return error

See merge request fsel/api/fsel_src!36

+++ Services/Fsel.Identity/Fsel.Identity.Domain/Models/EntityModels/SearchStudentsInClassModel.cs
@@ -11,6 +11,6 @@ namespace Fsel.Identity.Domain.Models.EntityModels
         public string? Code { get; set; }
         public string? Email { get; set; }
         public DateTime? BirthDay { get; set; }
-        public long PTPoint { get; set; }
+        public double PTPoint { get; set; }
     }
 }
